remove cost calculator
