+++ src/flows/CostCalculator/index.ts
@@ -3,8 +3,7 @@ export { CostCalculatorForm } from './CostCalculatorForm';
 export { CostCalculatorSubmitButton } from './CostCalculatorSubmitButton';
 export { CostCalculatorResetButton } from './CostCalculatorResetButton';
 export { CostCalculatorDisclaimer } from './CostCalculatorDisclaimer';
-export { useCostCalculator } from './hooks';
-export { useCostCalculatorEstimationPdf } from './api';
+export { useCostCalculator, useCostCalculatorEstimationPdf } from './hooks';
 export { disclaimerData } from './disclaimerUtils';
 export { CostCalculatorResults } from './Results/CostCalculatorResults';
 export { buildPayload as buildCostCalculatorEstimationPayload } from './utils';
